Added Resources Class, Fixed Tower
